Compiler Errors Prevents Program From Running
Error present in trying to access the CMake Debug file.
Project is unable to continue compiling. Before the issue,
program was edited to try and parse through a text file to
load a map, but was running into errors. This will likely be
the last commit before making a new copy of the final
project repository.