Implement User Movement Logic
Create Character class to track the display image of the current
character as well as the location. Character is bound by the frame
of the window and cannot pass through tiles that are categorized as
solid. Character will turn based on the direction that was input.
All sprites upon load-in are correct and up to date, but additional
features revealed after interactions will still need to be
implemented.