Allow Character to Navigate Between Scenes
Implement a "TEL" dynamic type for the scenes, which stands for
Teleport. Based on this logic, the character can navigate between
the scenes to travel back and forth. Character must be standing on a
Teleport tile and press the corresponding key leading to the side of
the map that it extends (ex. right on scene one teleport spot, left
on scene two teleport spot) in order to teleport.