Implement User-Map Interaction Logic
Create interactive actions depending on whether tiles are
dynamic or not. Track user interaction with the tiles and record the
state of the playing field between scene changes. Create Item
class that will represent the objects that can be collected by
the character. Implement certain special conditions such as
requiring the user to collect bees to defeat minions or not being
able to collect any more items if the inventory is full. Organize
area below the Playing Field to track information for the user,
such as descriptions, inventory and stats.