Implement Game Introduction Page and Win State
By utilizing enums, game states can be recorded and updated
as the application is running. Upon starting, application will show
the Introduction screen, which explains the objective, game content
and the mechanics of the game. Application will show the win screen
with a message upon interaction with the goal tile (in this case, it
is the poro).

Note: Lose states and final boss have not yet been implemented, but
they are in progress.